serverless/functions/audio2srt/requirements.txt
boto3
pydub
google-cloud-speech
google-cloud-storage
srt
# webvtt-py